Consolidate duplicated and orphaned test classes in bosk-core and bosk-logback.
Merge the two ReferenceUtils.parameterType test classes into one, fold BoskTest
into BoskConstructorTest, and fold RecordingTurboFilterCaptureTest into
RecordingTurboFilterTest. No test scenarios are lost.
